Add files via upload
It's a security review report of the puppyRaffle smart contract which is a raffle of puppy NFT where players can participate and whoever wins the lottery receives money and Puppy NFT according to rarity. I got some exploits in the contract. you can see it for more information about the report and PuppyRaffle contract (codehawks).